Some issues in the core identified by findbugs

+++ core/src/main/java/org/apache/cxf/feature/FastInfosetFeature.java
@@ -59,24 +59,24 @@ public class FastInfosetFeature extends AbstractFeature {
 
         FIStaxOutInterceptor out = new FIStaxOutInterceptor(force);
         if (serializerAttributeValueMapMemoryLimit != null && serializerAttributeValueMapMemoryLimit.intValue() > 0) {
-            out.setSerializerAttributeValueMapMemoryLimit(serializerAttributeValueMapMemoryLimit.intValue());
+            out.setSerializerAttributeValueMapMemoryLimit(serializerAttributeValueMapMemoryLimit);
         }
         if (serializerMinAttributeValueSize != null && serializerMinAttributeValueSize.intValue() > 0) {
-            out.setSerializerMinAttributeValueSize(serializerMinAttributeValueSize.intValue());
+            out.setSerializerMinAttributeValueSize(serializerMinAttributeValueSize);
         }
         if (serializerMaxAttributeValueSize != null && serializerMaxAttributeValueSize.intValue() > 0) {
-            out.setSerializerMaxAttributeValueSize(serializerMaxAttributeValueSize.intValue());
+            out.setSerializerMaxAttributeValueSize(serializerMaxAttributeValueSize);
         }
         if (serializerCharacterContentChunkMapMemoryLimit != null
                 && serializerCharacterContentChunkMapMemoryLimit.intValue() > 0) {
             out.setSerializerCharacterContentChunkMapMemoryLimit(
-                    serializerCharacterContentChunkMapMemoryLimit.intValue());
+                    serializerCharacterContentChunkMapMemoryLimit);
         }
         if (serializerMinCharacterContentChunkSize != null && serializerMinCharacterContentChunkSize.intValue() > 0) {
-            out.setSerializerMinCharacterContentChunkSize(serializerMinCharacterContentChunkSize.intValue());
+            out.setSerializerMinCharacterContentChunkSize(serializerMinCharacterContentChunkSize);
         }
         if (serializerMaxCharacterContentChunkSize != null && serializerMaxCharacterContentChunkSize.intValue() > 0) {
-            out.setSerializerMaxCharacterContentChunkSize(serializerMaxCharacterContentChunkSize.intValue());
+            out.setSerializerMaxCharacterContentChunkSize(serializerMaxCharacterContentChunkSize);
         }
         
         provider.getInInterceptors().add(in);
